pyproject.toml
[tool.poetry]
name = "xiv-static-tracker"
version = "0.1.0"
description = ""
authors = ["ishamyyl <ross.hodapp@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
starlette = "^0.31.1"
uvicorn = {version = "^0.30", extras = ["standard"]}
jinja2 = "^3.1.2"
python-multipart = "^0.0.6"
itsdangerous = "^2.1.2"
tortoise-orm = "^0.20.0"
asyncpg = "^0.29.0"
aerich = "^0.7.2"

[tool.pyright]
venvPath = "/home/zen/.cache/pypoetry/virtualenvs/"
venv = "xiv-static-tracker-BEsc8ryA-py3.11"

[tool.aerich]
tortoise_orm = "main.TORTOISE_CONFIG"
location = "./migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
